% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMI510.R
\name{adj_R2}
\alias{adj_R2}
\title{adj_R2
Calculate the adjusted r-squared statistic between predicted and ground truth continuous variables}
\usage{
adj_R2(pred, truth, n_p)
}
\arguments{
\item{truth, }{n_p}
}
\value{
the adjusted r-squared statistic between predicted and ground truth continuous variables
}
\description{
adj_R2
Calculate the adjusted r-squared statistic between predicted and ground truth continuous variables
}
