% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMI510.R
\name{is_max}
\alias{is_max}
\title{is_max
where x equals its maximum value}
\usage{
is_max(x, na.rm = T)
}
\arguments{
\item{na.rm}{}
}
\value{
a logical with TRUE
}
\description{
is_max
where x equals its maximum value
}
