% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMI510.R
\name{rando}
\alias{rando}
\title{rando
tests whether x is an atomic vector or dataframe-like object}
\usage{
rando(x, n = 1, replace = T)
}
\arguments{
\item{n, }{replace}
}
\value{
n samples or n rows
}
\description{
rando
tests whether x is an atomic vector or dataframe-like object
}
