% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMI510.R
\name{recall}
\alias{recall}
\title{recall
Calculate recall based on comparing predicted and training labels}
\usage{
recall(pred, truth)
}
\arguments{
\item{truth}{}
}
\value{
recall
}
\description{
recall
Calculate recall based on comparing predicted and training labels
}
