% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMI510.R
\name{sensitivity}
\alias{sensitivity}
\title{sensitivity
Calculate sensitivity based on comparing predicted and training labels}
\usage{
sensitivity(pred, truth)
}
\arguments{
\item{truth}{}
}
\value{
sensitivity
}
\description{
sensitivity
Calculate sensitivity based on comparing predicted and training labels
}
