% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMI510.R
\name{specificity}
\alias{specificity}
\title{specificity
Calculate specificity based on comparing predicted and training labels}
\usage{
specificity(pred, truth)
}
\arguments{
\item{truth}{}
}
\value{
specificity
}
\description{
specificity
Calculate specificity based on comparing predicted and training labels
}
